2021/day02/test_aoc_202102.py
# test_aoc_202102.py

import pathlib
import pytest
import pandas as pd
import aoc_202102 as aoc

PUZZLE_DIR = pathlib.Path(__file__).parent

@pytest.fixture
def example1():
    puzzle_input = (PUZZLE_DIR / "example1.txt")
    return aoc.parse(puzzle_input)

@pytest.fixture
def example2():
    puzzle_input = (PUZZLE_DIR / "example1.txt")
    return aoc.parse(puzzle_input)

# @pytest.mark.skip(reason="Not implemented")
def test_parse_example1(example1):
    """Test that input is parsed properly"""
    df1 = pd.DataFrame({'position': ['forward', 'down', 'forward','up', 'down','forward'], 'units': [5, 5, 8, 3, 8, 2]})
    pd.testing.assert_frame_equal(example1, df1)


# pytest.mark.skip(reason="Not implemented")
def test_part1_example1(example1):
    """Test part 1 on example input"""
    assert aoc.part1(example1) == 150

# @pytest.mark.skip(reason="Not implemented")
def test_part2_example2(example2):
    """Test part 2 on example input"""
    assert aoc.part2(example2) == 900